requirements.txt
numpy
scipy
librosa
pydub
webrtcvad
pyannote.audio
transformers
torch
pytest
ffmpeg-python